Move reusable components from src/app into src/appComponents
