+++ apps/store/src/app/[locale]/not-found.tsx
@@ -1,4 +1,4 @@
-import { NotFoundPageContent } from './NotFoundPageContent'
+import { NotFoundPageContent } from '@/appComponents/RootLayout/NotFoundPageContent'
 
 export default function NotFound() {
   return <NotFoundPageContent />
