+++ apps/store/tsconfig.json
@@ -5,6 +5,9 @@
     "outDir": "dist",
     "baseUrl": "src",
     "paths": {
+      "@/appComponents/*": [
+        "appComponents/*"
+      ],
       "@/lib/*": [
         "lib/*"
       ],
